requirements.txt
huggingface_hub
python-dotenv
PyYAML
torch
transformers
optimum
auto-gptq
bitsandbytes